+++ tempodb/encoding/vparquet5/block_traceql_test.go
@@ -369,6 +369,7 @@ func TestBackendNilValueBlockSearchTraceQL(t *testing.T) {
 					{
 						ScopeSpans: []ScopeSpans{
 							{
+								SpanCount: 2,
 								Spans: []Span{
 									{
 										// this span has nil values for everything
@@ -418,6 +419,7 @@ func TestBackendNilValueBlockSearchTraceQL(t *testing.T) {
 						},
 						ScopeSpans: []ScopeSpans{
 							{
+								SpanCount: 1,
 								Spans: []Span{
 									{
 										SpanID: []byte("nil-test-span-2"),
@@ -1034,6 +1036,7 @@ func fullyPopulatedTestTraceWithOption(id common.ID, parentIDTest bool) *Trace {
 								attr("scope-attr-bool", true),
 							},
 						},
+						SpanCount: 1,
 						Spans: []Span{
 							{
 								SpanID:                 []byte("spanid"),
@@ -1138,6 +1141,7 @@ func fullyPopulatedTestTraceWithOption(id common.ID, parentIDTest bool) *Trace {
 								attr("scope-attr-str", "scope-attr-2"),
 							},
 						},
+						SpanCount: 1,
 						Spans: []Span{
 							{
 								SpanID:                 []byte("spanid2"),